fix: merge전 코드 수정
