+++ hotsix-react-app/App/src/screens/Group/GroupNoticeScreen.js
@@ -4,6 +4,8 @@ import axios from 'axios';
 
 const GroupNoticeScreen = ({ navigation }) => {
   const [notices, setNotices] = useState([]);
+  const [title, setTitle] = useState('');
+  const [content, setContent] = useState('');
 
   useEffect(() => {
     loadNotices();
@@ -13,6 +15,18 @@ const GroupNoticeScreen = ({ navigation }) => {
     axios.get('http://192.168.0.120:3001/notices')
       .then(response => {
         setNotices(response.data);
+      })
+      .catch(error => {
+        console.error(error);
+      });
+  };
+
+  const createNotice = () => {
+    axios.post('http://192.168.0.120:3001/notices', { title, content })
+      .then(response => {
+        setTitle('');
+        setContent('');
+        loadNotices();
       })
       .catch(error => {
         console.error(error);
@@ -36,14 +50,22 @@ const GroupNoticeScreen = ({ navigation }) => {
     </TouchableOpacity>
   );
 
-  const navigateToCreateNotice = () => {
-    navigation.navigate('GroupNoticeCreation');
-  };
-
   return (
     <View style={styles.container}>
       <View style={styles.inputContainer}>
-        <Button title="작성" onPress={navigateToCreateNotice} />
+        <TextInput
+          style={styles.input}
+          placeholder="제목"
+          value={title}
+          onChangeText={text => setTitle(text)}
+        />
+        <TextInput
+          style={styles.input}
+          placeholder="내용"
+          value={content}
+          onChangeText={text => setContent(text)}
+        />
+        <Button title="작성" onPress={createNotice} />
       </View>
       <FlatList
         data={notices}
@@ -53,3 +75,30 @@ const GroupNoticeScreen = ({ navigation }) => {
     </View>
   );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+  },
+  inputContainer: {
+    marginBottom: 16,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    borderRadius: 4,
+    padding: 8,
+    marginBottom: 8,
+  },
+  noticeTitle: {
+    fontSize: 16,
+  },
+  noticeContent: {
+    borderBottomWidth: 1,
+    borderBottomColor: '#ccc',
+    marginVertical: 8,
+  },
+});
+
+export default GroupNoticeScreen;
